requirements.txt
asgiref==3.2.7
boto3==1.13.6
botocore==1.16.6
Django==3.0.6
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.5
numpy==1.18.4
pandas==1.0.3
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
